An incomplete UCI protocol. Have not had enough time recently to finish, so I'm just getting a commit out eventually.

+++ src/engine/search_limits.rs
@@ -1,13 +1,14 @@
 use std::time::Instant;
 
+#[derive(Copy, Clone)]
 pub struct SearchLimits{
     timer: Instant,
-    soft_stop: u128,
-    hard_stop: u128,
+    soft_stop: u32,
+    hard_stop: u32,
 }
 
 impl SearchLimits {
-    pub fn new(soft_stop: u128, hard_stop: u128) -> SearchLimits {
+    pub fn new(soft_stop: u32, hard_stop: u32) -> SearchLimits {
         SearchLimits{
             timer: Instant::now(),
             soft_stop,
@@ -17,11 +18,11 @@ impl SearchLimits {
 
 
     pub fn is_soft_stop(&self) -> bool{
-        self.soft_stop <= self.timer.elapsed().as_millis()
+        self.soft_stop <= self.timer.elapsed().as_millis() as u32
     }
 
     pub fn hard_stop(&self) -> bool{
-        self.hard_stop <= self.timer.elapsed().as_millis()
+        self.hard_stop <= self.timer.elapsed().as_millis() as u32
     }
     
     pub fn ms_elapsed(&self) -> u128{
